requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
plotly>=5.18.0
dash>=2.0.0
dash-bootstrap-components>=0.13.1
nltk>=3.6.2
joblib>=1.0.1
dataclasses>=0.6
web3>=5.24.0
eth-account>=0.5.6
opencv-python>=4.5.3
face-recognition>=1.3.0
requests>=2.26.0
pytesseract>=0.3.8
Pillow>=8.3.1
cryptography>=3.4.7
transformers>=4.35.0
torch>=2.1.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
spacy>=3.7.2
prophet>=1.1.0
statsmodels>=0.13.0
textblob>=0.15.3
wordcloud>=1.8.1
folium>=0.12.1
networkx>=2.6.3
pycountry>=22.3.5
scipy>=1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
aiohttp>=3.8.0
psutil>=5.8.0
PyYAML>=5.4.1
Jinja2>=3.1.2
schedule>=1.1.0
pytz>=2021.1
websockets>=10.0
PyJWT>=2.1.0
python-dateutil>=2.8.2
pydantic>=1.8.2
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
email-validator>=1.1.3
pdf2image>=1.16.0
PyMuPDF>=1.19.0
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.2.0/en_core_web_trf-3.2.0.tar.gz
camembert-base @ https://huggingface.co/camembert-base/resolve/main/pytorch_model.bin
xgboost>=1.5.0
selenium>=4.1.0
aiosmtplib>=1.1.6
webdriver_manager>=3.5.2
lxml>=4.9.0
html5lib>=1.1
tqdm>=4.62.3
kaleido>=0.2.1
pandas>=2.1.0
numpy>=1.24.0
pdfkit>=1.0.0
wkhtmltopdf>=0.2
reportlab>=3.6.12
weasyprint>=54.3
python-jose[cryptography]>=3.3.0
python-jwt>=3.3.0
bcrypt>=4.0.1
pyotp>=2.8.0
qrcode>=7.3
pillow>=8.3.1
redis>=4.5.1
rate-limit>=2.2.1
aioredis>=2.0.1
python-telegram-bot>=13.7
twilio>=7.0.0
slack-sdk>=3.19.5
firebase-admin>=5.3.0
newspaper3k>=0.2.8
tweepy>=4.12.0
linkedin-api>=0.1.7
python-slugify>=8.0.1
markdown>=3.5.1
bleach>=6.1.0
beautifulsoup4>=4.12.2 